Update patch set 1
Patch Set 1:

Hi Andrew, x86 target may have some issues with SSE.
For x86, it may have issues due to unaligned stack element.
Actually, we are using x86-atom internally, so that should be not problem with x86-atom and I propose not to do the change,
Any comments are welcome.

Patch-set: 1