Update patch set 1
Patch Set 1:

I'm all for adding optimizations as well.

It looks like the assembly file is almost an exact copy of this.  In libjpeg-turbo, it is in a subdirectory.
https://github.com/libjpeg-turbo/libjpeg-turbo/blob/master/simd/jsimd_arm64_neon.S

A diff between libjpeg-turbo and the file in this CL show that this CL uses INT32 where the latest turbo uses JLONG.  Is there a reason for this?  Or is it just because this is pulled from an old commit and turbo has updated since then?  I haven't had any luck tracking down the commit hash you've referenced either.

My main concerns are regarding adding another fork to this copy of libjpeg.  The first fork has already caused correctness bugs and complications with updating.  libjpeg and libjpeg-turbo are similar, but not identical.  I'm guessing that the patches to jdcolor.c and jddctmgr.c are manual?  Which may be fine given that they are not extremely complicated.  But it's worth noting that the patch to jdcolor.c actually builds on the existing ANDROID_RGB fork by using JCS_8888 (turbo has JCS_EXT_RGBA and JCS_EXT_BGRA).

Can you describe what sort of testing you've performed?  I think it's important that we test comprehensively, given that we are moving farther away from the standard copy of the library.

Patch-set: 1