fix crash when compiled with Clang
This patch fixes https://github.com/android-ndk/ndk/issues/110
The crash seemed to come from a change in LLVM and was reported to the LLVM
bugzilla https://llvm.org/bugs/show_bug.cgi?id=28393

Upon investigation, LLVM optimized away the zeroing of the upper bits of the x2
register before calling color_convert.  The convention is for the callee to do
any truncation needed.  The patch adds a zero-extend for the low word w2 in x2
to follow the declaration of the third argument in color_convert: INPUT_ROW of
type JDIMENSION should be an unsigned int, so the upper part of x2 should be
zeroed out.

Change-Id: Ie2fc005914a956ab172362e3abecd764fe6ab861